Move event storage from tako to HyperQueue
