+++ crates/hyperqueue/src/lib.rs
@@ -8,6 +8,7 @@ pub mod stream;
 pub mod transfer;
 pub mod worker;
 
+pub mod events;
 #[cfg(test)]
 pub(crate) mod tests;
