+++ crates/tako/src/server/mod.rs
@@ -3,8 +3,6 @@ pub use start::server_start;
 pub mod client;
 pub mod comm;
 pub mod core;
-pub mod gateway;
-pub mod monitoring;
 pub mod reactor;
 pub mod rpc;
 mod start;
